requirements.txt
django
gunicorn
django-heroku
vk_api
pillow
